feat: allow for all tiles to connect to Donobrick
While this should've been a feature since the start, I suspect that we confused tileMergeDirt with tileBlendAll. They are different!

One indicates that the tile, itself, has textures to connect to dirt (and possibly dirt-alikes).
Tother makes it so that OTHER tiles CAN connect to our modded tile. Essentially making it behave like vanilla brick and brick-alike tiles.